+++ src/Mod/Spreadsheet/App/Sheet.cpp
@@ -35,6 +35,7 @@
 #include <App/Document.h>
 #include <App/DynamicProperty.h>
 #include <App/FeaturePythonPyImp.h>
+#include <App/ExpressionParser.h>
 #include <Base/Exception.h>
 #include <Base/FileInfo.h>
 #include <Base/Placement.h>
