+++ src/Mod/Spreadsheet/Gui/PropertiesDialog.cpp
@@ -26,6 +26,7 @@
 #include <Base/Tools.h>
 #include <App/Range.h>
 #include <App/Document.h>
+#include <App/ExpressionParser.h>
 #include <Gui/Command.h>
 #include "ui_PropertiesDialog.h"
